Include acceptance behaviours into acceptance tests
